requirements.txt
# Core dependencies
requests
trafilatura

# RAG and NLP
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
transformers>=4.30.0
torch>=2.0.0

# Translation models
sentencepiece>=0.1.99
protobuf>=3.20.0

# OpenAI (updated to v1.0+)
openai>=1.0.0

# Evaluation
nltk>=3.8.0
numpy>=1.24.0

# Web app
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0